Test CheckPaginationOnJobPage updated
