+++ ui_tests/PlaywrightAutomation/PlaywrightAutomation.csproj
@@ -33,6 +33,8 @@
     <None Update="appsettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <Folder Include="SpecFlowTests\CareerPage\ApplyForAJob\" />
+    <Folder Include="SpecFlowTests\CareerPage\JobBlock\" />
   </ItemGroup>
 
 </Project>
